requirements.txt
Flask
pymongo
torch
transformers
dnspython
tld
numpy<2
scikit-learn
flask-cors
python-dotenv